Fix data point type

+++ data-point.go
@@ -35,7 +35,7 @@ type DataPoint struct {
 	TemperatureMin             float64 `json:"temperatureMin"`
 	TemperatureMinTime         float64 `json:"temperatureMinTime"`
 	Time                       Time    `json:"time"`
-	UVIndex                    int     `json:"uvIndex"`
+	UVIndex                    float64 `json:"uvIndex"`
 	UVIndexTime                *Time   `json:"uvIndexTime,omitempty"`
 	Visibility                 float64 `json:"visibility"`
 	WindBearing                float64 `json:"windBearing"`
